API testing and finalizing part 1.

+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\RoleResource;
 use App\Role;
+use DB;
+use Gate;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -17,6 +19,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('view', 'roles');
+
         $roles = Role::all();
         return RoleResource::collection($roles);
     }
@@ -29,13 +33,15 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::create($request->only('name'));
 
         if ($permissions = $request->input('permissions')) {
             foreach ($permissions as $permission_id) {
-                \DB::table('role_permission')->insert([
+                DB::table('role_permission')->insert([
                     'role_id' => $role->id,
-                    'permission_id'=> $permission_id,
+                    'permission_id' => $permission_id,
                 ]);
             }
         }
@@ -51,6 +57,8 @@ class RoleController extends Controller
      */
     public function show($id)
     {
+        Gate::authorize('view', 'roles');
+
         return new RoleResource(Role::find($id));
     }
 
@@ -63,16 +71,18 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::findOrFail($id);
         $role->update($request->only('name'));
 
-        \DB::table('role_permission')->where('role_id', $role->id)->delete();
+        DB::table('role_permission')->where('role_id', $role->id)->delete();
 
         if ($permissions = $request->input('permissions')) {
             foreach ($permissions as $permission_id) {
-                \DB::table('role_permission')->insert([
+                DB::table('role_permission')->insert([
                     'role_id' => $role->id,
-                    'permission_id'=> $permission_id,
+                    'permission_id' => $permission_id,
                 ]);
             }
         }
@@ -88,7 +98,9 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
-        \DB::table('role_permission')->where('role_id', $id)->delete();
+        Gate::authorize('edit', 'roles');
+
+        DB::table('role_permission')->where('role_id', $id)->delete();
 
         Role::destroy($id);
